feat: star ratings, timer ,and new style
Feat the star ratings and timer which limit the player in specific time and limited number of movements, When a certain limit is reached the game stops and popup message appears.
If the cards not matched, there color change to red and shakes for a while before flipped.
In summery, these are the changes :
- Win and lost popup message.
- Timer one minute.
- Stars decrease with movements.
- Restart bottom to reset the game board.
- Match function, no match function.
- New colors.